// src/types/contentful.d.ts
// THIS FILE IS AUTOMATICALLY GENERATED. DO NOT MODIFY IT.

import { Document } from "@contentful/rich-text-types";
import { Asset, Entry } from "contentful";

export interface IEventFields {
	/** Title */
	title: string;

	/** Description */
	description?: string | undefined;

	/** Date */
	date?: string | undefined;

	/** Link To Event */
	linkToEvent?: string | undefined;

	/** Banner */
	banner: Asset;
}

export interface IEvent extends Entry<IEventFields> {
	sys: {
		id: string;
		type: string;
		createdAt: string;
		updatedAt: string;
		locale: string;
		contentType: {
			sys: {
				id: "event";
				linkType: "ContentType";
				type: "Link";
			};
		};
	};
}

export interface IProjectFields {
	/** Hero Banner */
	heroBanner?: Asset | undefined;

	/** Title */
	title: string;

	/** Description */
	description?: string | undefined;

	/** Location */
	location?: string | undefined;

	/** Project Content */
	projectContent: Document;

	/** Event Date */
	eventDate: string;

	/** Project */
	project?: string | undefined;

	/** NGO/ Partnerships */
	ngoPartnerships?: string | undefined;

	/** SDG */
	sdg?: string | undefined;

	/** Start date */
	startDate?: string | undefined;

	/** End date */
	endDate?: string | undefined;
}

/** Setup a new project */

export interface IProject extends Entry<IProjectFields> {
	sys: {
		id: string;
		type: string;
		createdAt: string;
		updatedAt: string;
		locale: string;
		contentType: {
			sys: {
				id: "project";
				linkType: "ContentType";
				type: "Link";
			};
		};
	};
}

export type CONTENT_TYPE = "event" | "project";

export type IEntry = IEvent | IProject;

export type LOCALE_CODE = "en-US";

export type CONTENTFUL_DEFAULT_LOCALE_CODE = "en-US";